refactor: split test_tables_schema by taxonomy per test.instructions.md
Pure unit tests → tests/test_core/test_delta_client.py
Utils tests     → tests/test_utils/test_fab_cmd_table_utils.py
Integration     → tests/test_commands/test_tables_schema.py (only)
